+++ jopa-integration-tests-jena/src/test/java/cz/cvut/kbss/jopa/test/environment/JenaDataAccessor.java
@@ -1,16 +1,14 @@
 /**
  * Copyright (C) 2019 Czech Technical University in Prague
- *
- * This program is free software: you can redistribute it and/or modify it under
- * the terms of the GNU General Public License as published by the Free Software
- * Foundation, either version 3 of the License, or (at your option) any
- * later version.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
- * details. You should have received a copy of the GNU General Public License
- * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ * <p>
+ * This program is free software: you can redistribute it and/or modify it under the terms of the GNU General Public
+ * License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later
+ * version.
+ * <p>
+ * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied
+ * warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
+ * details. You should have received a copy of the GNU General Public License along with this program. If not, see
+ * <http://www.gnu.org/licenses/>.
  */
 package cz.cvut.kbss.jopa.test.environment;
 
@@ -22,6 +20,7 @@ import org.apache.jena.rdf.model.*;
 import java.net.URI;
 import java.util.ArrayList;
 import java.util.Collection;
+import java.util.Collections;
 import java.util.List;
 
 import static org.apache.jena.rdf.model.ResourceFactory.*;
@@ -31,54 +30,43 @@ import static org.junit.Assert.assertTrue;
 public class JenaDataAccessor implements DataAccessor {
 
     @Override
-    public void persistTestData(Collection<Triple> data, EntityManager em) throws Exception {
+    public void persistTestData(Collection<Quad> data, EntityManager em) throws Exception {
         final StorageConnector ds = em.unwrap(StorageConnector.class);
         ds.begin();
         final List<Statement> toAdd = new ArrayList<>(data.size());
-        for (Triple t : data) {
+        for (Quad t : data) {
             final Resource subject = createResource(t.getSubject().toString());
             final Property property = createProperty(t.getProperty().toString());
-            final RDFNode value;
-            if (t.getValue() instanceof URI) {
-                value = createResource(t.getValue().toString());
-            } else {
-                value = t.getValue() instanceof String ?
-                        ResourceFactory.createLangLiteral(t.getValue().toString(), t.getLanguage()) :
-                        ResourceFactory.createTypedLiteral(t.getValue());
-            }
-            toAdd.add(createStatement(subject, property, value));
+            final RDFNode value = toValue(t.getValue(), t.getLanguage());
+            ds.add(Collections.singletonList(createStatement(subject, property, value)),
+                    t.getContext() != null ? t.getContext().toString() : null);
         }
         ds.add(toAdd, null);
         ds.commit();
     }
 
-    @Override
-    public void verifyDataPresent(Collection<Triple> expected, EntityManager em) {
-        final Dataset ds = em.unwrap(Dataset.class);
-        final Model model = ds.getDefaultModel();
-        expected.forEach(t -> assertTrue(doesTripleExist(t, model)));
-    }
-
-    private boolean doesTripleExist(Triple triple, Model model) {
-        if (triple.getValue() instanceof URI) {
-            return model.contains(createResource(triple.getSubject().toString()),
-                    createProperty(triple.getProperty().toString()), createResource(triple.getValue().toString()));
-        } else {
-            if (triple.getLanguage() != null) {
-                return model.contains(createResource(triple.getSubject().toString()),
-                        createProperty(triple.getProperty().toString()),
-                        createLangLiteral(triple.getValue().toString(), triple.getLanguage()));
-            } else {
-                return model.contains(createResource(triple.getSubject().toString()),
-                        createProperty(triple.getProperty().toString()), createTypedLiteral(triple.getValue()));
-            }
-        }
+    private RDFNode toValue(Object value, String language) {
+        return value instanceof URI ? createResource(value.toString()) :
+               (value instanceof String ? ResourceFactory.createLangLiteral(value.toString(), language) :
+                ResourceFactory.createTypedLiteral(value));
     }
 
     @Override
-    public void verifyDataNotPresent(Collection<Triple> data, EntityManager em) {
+    public void verifyDataPresent(Collection<Quad> expected, EntityManager em) {
         final Dataset ds = em.unwrap(Dataset.class);
-        final Model model = ds.getDefaultModel();
-        data.forEach(t -> assertFalse(doesTripleExist(t, model)));
+        expected.forEach(t -> assertTrue(doesQuadExist(t,
+                t.getContext() == null ? ds.getDefaultModel() : ds.getNamedModel(t.getContext().toString()))));
+    }
+
+    private boolean doesQuadExist(Quad quad, Model model) {
+        return model.contains(createResource(quad.getSubject().toString()),
+                createProperty(quad.getProperty().toString()), toValue(quad.getValue(), quad.getLanguage()));
+    }
+
+    @Override
+    public void verifyDataNotPresent(Collection<Quad> data, EntityManager em) {
+        final Dataset ds = em.unwrap(Dataset.class);
+        data.forEach(t -> assertFalse(doesQuadExist(t,
+                t.getContext() == null ? ds.getDefaultModel() : ds.getNamedModel(t.getContext().toString()))));
     }
 }